Move HTTPartyWrapper to AdRoll namespace; add spec tests for normalizer_proc
