+++ lib/adroll.rb
@@ -29,7 +29,7 @@ require 'adroll/uhura/attributions'
 require 'adroll/uhura/deliveries'
 require 'adroll/uhura/userlists'
 
-require 'httparty_wrapper/httparty_wrapper'
+require 'adroll/httparty_wrapper'
 
 module AdRoll
   def self.api_base_url
